restructure the directory structure
